+++ src/allocs/c_alloc.rs
@@ -1,22 +1,23 @@
 use {
     crate::{
-        error::Error,
-        ffi::c_alloc::{c_alloc, c_dealloc, c_zalloc, grow_aligned, shrink_aligned},
-        helpers::null_q_dyn_zsl_check,
+        error::{Cause, Error},
+        ffi::c_alloc::{c_alloc, c_dealloc, c_zalloc, grow_aligned, shrink_aligned}
+        ,
         layout::Layout,
         traits::{
-            AllocError,
-            alloc::{Alloc, Dealloc, Grow, Realloc, Shrink}
+            alloc::{Alloc, Dealloc, Grow, Realloc, Shrink},
+            AllocError
         }
     },
     ::core::{
         cmp::{Ord, Ordering},
         ffi::c_void,
+        ops::Fn,
         ptr::NonNull,
         result::Result::{self, Err, Ok}
-    }
+    },
+    ::cty::c_int
 };
-// TODO: we should use the builtin malloc and realloc if align <= guaranteed align
 
 macro_rules! sz_check {
     ($err:ident, $old:ident $cmp:tt $new:ident) => {
@@ -26,17 +27,38 @@ macro_rules! sz_check {
     };
 }
 
-#[cfg_attr(miri, track_caller)]
-fn pad_then_alloc(
+fn null_q_dyn_zsl_check_or_errcode<F: Fn(Layout) -> (*mut c_void, c_int)>(
     layout: Layout,
-    alloc: unsafe fn(usize, usize) -> *mut c_void
+    f: F
 ) -> Result<NonNull<u8>, Error> {
+    if layout.is_zero_sized() {
+        Err(Error::ZeroSizedLayout)
+    } else {
+        let (ptr, status) = f(layout);
+        if status == -1 {
+            // SAFETY: errcode of -1 means it's already been checked and is guaranteed success
+            Ok(unsafe { NonNull::new_unchecked(ptr.cast()) })
+        } else if status == 0 {
+            null_q_dyn(ptr, layout)
+        } else {
+            #[cfg(feature = "os_err_reporting")]
+            {
+                Err(Error::AllocFailed(layout, Cause::OSErr(status as c_int)))
+            }
+            #[cfg(not(feature = "os_err_reporting"))]
+            {
+                Err(Error::AllocFailed(layout, Cause::Unknown))
+            }
+        }
+    }
+}
+
+#[cfg_attr(miri, track_caller)]
+fn pad_then_alloc(layout: Layout, zeroed: bool) -> Result<NonNull<u8>, Error> {
     let padded = tri!(do layout.to_posix_memalign_compatible());
-    null_q_dyn_zsl_check(
-        layout,
-        // SAFETY: we rounded up the layout's values to satisfy the requirements.
-        |_| unsafe { alloc(padded.align(), padded.size()) }
-    )
+    let fun = if zeroed { c_zalloc } else { c_alloc };
+    // SAFETY: padded layout meets the function's requirements
+    null_q_dyn_zsl_check_or_errcode(layout, |_| unsafe { fun(padded.align(), padded.size()) })
 }
 
 #[cfg_attr(miri, track_caller)]
@@ -44,20 +66,21 @@ unsafe fn pad_then_grow(
     ptr: NonNull<u8>,
     old_layout: Layout,
     new_layout: Layout,
-    alloc: unsafe fn(usize, usize) -> *mut c_void
+    zeroed: bool
 ) -> Result<NonNull<u8>, Error> {
     sz_check!(GrowSmallerNewLayout, old_layout > new_layout);
 
     let old_padded = tri!(do old_layout.to_posix_memalign_compatible());
     let new_padded = tri!(do new_layout.to_posix_memalign_compatible());
 
-    null_q_dyn_zsl_check(new_layout, |_| {
+    null_q_dyn_zsl_check_or_errcode(new_layout, |_| {
         grow_aligned(
             ptr.as_ptr().cast(),
+            old_padded.align(),
             old_padded.size(),
             new_padded.align(),
             new_padded.size(),
-            alloc
+            zeroed
         )
     })
 }
@@ -67,12 +90,12 @@ unsafe fn pad_then_realloc(
     ptr: NonNull<u8>,
     old_layout: Layout,
     new_layout: Layout,
-    alloc: unsafe fn(usize, usize) -> *mut c_void
+    zeroed: bool
 ) -> Result<NonNull<u8>, Error> {
     let old_padded = tri!(do old_layout.to_posix_memalign_compatible());
     let new_padded = tri!(do new_layout.to_posix_memalign_compatible());
 
-    null_q_dyn_zsl_check(new_layout, |_| {
+    null_q_dyn_zsl_check_or_errcode(new_layout, |_| {
         let old_ptr = ptr.as_ptr().cast();
         let old_size = old_padded.size();
         let old_align = old_padded.align();
@@ -83,18 +106,18 @@ unsafe fn pad_then_realloc(
         match old_size.cmp(&new_padded.size()) {
             // SAFETY: caller guarantees that `old_ptr` and `old_size` are valid, we just
             // checked that `size >= old_size`
-            Ordering::Less => unsafe { grow_aligned(old_ptr, old_size, align, size, alloc) },
+            Ordering::Less => unsafe { grow_aligned(old_ptr, old_align, old_size, align, size, zeroed) },
             Ordering::Equal => {
                 if align > old_align {
                     // SAFETY: above
-                    unsafe { grow_aligned(old_ptr, old_size, align, size, alloc) }
+                    unsafe { grow_aligned(old_ptr, old_align, old_size, align, size, zeroed) }
                 } else {
-                    old_ptr
+                    (old_ptr, -1)
                 }
             }
             // SAFETY: caller guarantees that `old_ptr` and `size` are valid, we just checked
             // that `size <= old_size`
-            Ordering::Greater => unsafe { shrink_aligned(old_ptr, align, size) }
+            Ordering::Greater => unsafe { shrink_aligned(old_ptr, old_align, align, size) }
         }
     })
 }
@@ -115,13 +138,13 @@ impl Alloc for CAlloc {
     #[cfg_attr(miri, track_caller)]
     #[inline]
     fn alloc(&self, layout: Layout) -> Result<NonNull<u8>, Error> {
-        pad_then_alloc(layout, c_alloc)
+        pad_then_alloc(layout, false)
     }
 
     #[cfg_attr(miri, track_caller)]
     #[inline]
     fn zalloc(&self, layout: Layout) -> Result<NonNull<u8>, Error> {
-        pad_then_alloc(layout, c_zalloc)
+        pad_then_alloc(layout, true)
     }
 }
 impl Dealloc for CAlloc {
@@ -133,7 +156,8 @@ impl Dealloc for CAlloc {
         } else if ptr == layout.dangling() {
             Err(Error::DanglingDeallocation)
         } else {
-            c_dealloc(ptr.as_ptr().cast());
+            let padded = tri!(do layout.to_posix_memalign_compatible());
+            c_dealloc(ptr.as_ptr().cast(), padded.align());
             Ok(())
         }
     }
@@ -146,7 +170,7 @@ impl Grow for CAlloc {
         old_layout: Layout,
         new_layout: Layout
     ) -> Result<NonNull<u8>, Error> {
-        pad_then_grow(ptr, old_layout, new_layout, c_alloc)
+        pad_then_grow(ptr, old_layout, new_layout, false)
     }
     #[cfg_attr(miri, track_caller)]
     unsafe fn zgrow(
@@ -155,7 +179,7 @@ impl Grow for CAlloc {
         old_layout: Layout,
         new_layout: Layout
     ) -> Result<NonNull<u8>, Error> {
-        pad_then_grow(ptr, old_layout, new_layout, c_zalloc)
+        pad_then_grow(ptr, old_layout, new_layout, true)
     }
 }
 impl Shrink for CAlloc {
@@ -168,10 +192,11 @@ impl Shrink for CAlloc {
     ) -> Result<NonNull<u8>, Error> {
         sz_check!(ShrinkLargerNewLayout, old_layout < new_layout);
 
+        let old_padded = tri!(do old_layout.to_posix_memalign_compatible());
         let new_padded = tri!(do new_layout.to_posix_memalign_compatible());
 
-        null_q_dyn_zsl_check(new_layout, |_| {
-            shrink_aligned(ptr.as_ptr().cast(), new_padded.align(), new_padded.size())
+        null_q_dyn_zsl_check_or_errcode(new_layout, |_| {
+            shrink_aligned(ptr.as_ptr().cast(), old_padded.align(), new_padded.align(), new_padded.size())
         })
     }
 }
@@ -183,7 +208,7 @@ impl Realloc for CAlloc {
         old_layout: Layout,
         new_layout: Layout
     ) -> Result<NonNull<u8>, Error> {
-        pad_then_realloc(ptr, old_layout, new_layout, c_alloc)
+        pad_then_realloc(ptr, old_layout, new_layout, false)
     }
     #[cfg_attr(miri, track_caller)]
     unsafe fn rezalloc(
@@ -192,8 +217,9 @@ impl Realloc for CAlloc {
         old_layout: Layout,
         new_layout: Layout
     ) -> Result<NonNull<u8>, Error> {
-        pad_then_realloc(ptr, old_layout, new_layout, c_zalloc)
+        pad_then_realloc(ptr, old_layout, new_layout, true)
     }
 }
 
 pub use crate::ffi::c_alloc as ffi;
+use crate::helpers::null_q_dyn;