+++ src/layout.rs
@@ -221,15 +221,11 @@ impl Layout {
     /// Creates a layout compatible with C's `posix_memalign` requirements from the given `size` and
     /// `align`.
     ///
-    /// C's `posix_memalign(alignment, size)` requires:
-    /// - `alignment` is a power of two, non-zero, and a multiple of <code>[size_of]::<*mut
-    ///   [c_void](::core::ffi::c_void)>()</code>.
-    /// - `size` is a multiple of `alignment`.
+    /// C's `posix_memalign(out, alignment, size)` requires `alignment` is a power of two, non-zero,
+    /// and a multiple of <code>[size_of]::<*mut [c_void](::core::ffi::c_void)>()</code>.
     ///
-    /// Therefore:
-    /// - `align` will be rounded up to the nearest multiple of <code>[size_of]::<*mut
-    ///   [c_void](::core::ffi::c_void)>()</code> if it isn't already.
-    /// - `size` will be rounded up to the nearest multiple of the resulting alignment.
+    /// Therefore, the alignment will be rounded up to the nearest multiple of
+    /// <code>[size_of]::<*mut [c_void](::core::ffi::c_void)>()</code> if it isn't already.
     ///
     /// This is semantically equivalent to <code>[Layout::from_size_align]\(size,
     /// align\).[and_then](Result::and_then)\(|l|
@@ -239,11 +235,8 @@ impl Layout {
     ///
     /// # Errors
     ///
-    /// - <code>Err([Error::InvalidLayout]\([self.align()](Layout::align),
-    ///   [self.align()](Layout::align), [LayoutErr::ZeroAlign]\))</code> if `align == 0`.
-    /// - <code>Err([Error::InvalidLayout]\([self.size()](Layout::size),
-    ///   [self.align()](Layout::align), [LayoutErr::CRoundUp]\))</code> if `size` rounded up to the
-    ///   new alignment would exceed [`USIZE_MAX_NO_HIGH_BIT`].
+    /// <code>Err([Error::InvalidLayout]\([self.align()](Layout::align),
+    /// [self.align()](Layout::align), [LayoutErr::ZeroAlign]\))</code> if `align == 0`.
     ///
     /// # Examples
     ///
@@ -252,17 +245,19 @@ impl Layout {
     /// let l = Layout::posix_memalign_compatible_from_size_align(10, 1).unwrap();
     ///
     /// assert!(l.align() >= usize::SZ);
-    /// assert_eq!(l.size() % l.align(), 0);
     /// assert!(l.size() >= 10);
-    /// // on 64-bit systems, l == Layout(size = 16, align = 8).
-    /// // 32-bit, l == Layout(size = 12, align = 4)
+    /// // on 64-bit systems, l == Layout(size = 10, align = 8).
+    /// // 32-bit, l == Layout(size = 10, align = 4)
     /// ```
     pub const fn posix_memalign_compatible_from_size_align(
         size: usize,
         align: usize
     ) -> Result<Layout, Error> {
+        if align == 0 {
+            return Err(Error::InvalidLayout(size, align, LayoutErr::ZeroAlign));
+        }
         let align_rounded = align_up(align, usize::SZ);
-        match Layout::from_size_align(align_up(size, align_rounded), align_rounded) {
+        match Layout::from_size_align(size, align_rounded) {
             Ok(l) => Ok(l),
             Err(_) => Err(Error::InvalidLayout(size, align, LayoutErr::CRoundUp))
         }
@@ -473,15 +468,11 @@ impl Layout {
 
     /// Converts this layout into one compatible with C's `posix_memalign` requirements.
     ///
-    /// C's `posix_memalign(alignment, size)` requires:
-    /// - `alignment` is a power of two, non-zero, and a multiple of <code>[size_of]::<*mut
-    ///   [c_void](::core::ffi::c_void)>()</code>.
-    /// - `size` is a multiple of `alignment`.
+    /// C's `posix_memalign(out, alignment, size)` requires `alignment` is a power of two, non-zero,
+    /// and a multiple of <code>[size_of]::<*mut [c_void](::core::ffi::c_void)>()</code>
     ///
-    /// Therefore:
-    /// - The alignment will be rounded up to the nearest multiple of <code>[size_of]::<*mut
-    ///   [c_void](::core::ffi::c_void)>()</code> if it isn't already.
-    /// - The size will be rounded up to the nearest multiple of the resulting alignment.
+    /// Therefore, the alignment will be rounded up to the nearest multiple of
+    /// <code>[size_of]::<*mut [c_void](::core::ffi::c_void)>()</code> if it isn't already.
     ///
     /// # Errors
     ///
@@ -491,7 +482,6 @@ impl Layout {
     /// - `align` is not a power of two.
     /// - `align` rounded up to <code>[size_of]::<*mut [c_void](::core::ffi::c_void)>()</code> would
     ///   exceed the maximum allowed alignment.
-    /// - `size` rounded up to the new alignment would exceed [`USIZE_MAX_NO_HIGH_BIT`].
     ///
     /// [size_of]: ::core::mem::size_of
     ///
@@ -503,17 +493,16 @@ impl Layout {
     /// let compatible = l.to_posix_memalign_compatible().unwrap();
     ///
     /// assert!(compatible.align() >= usize::SZ);
-    /// assert_eq!(compatible.size() % compatible.align(), 0);
     /// assert!(compatible.size() >= 10);
-    /// // on 64-bit systems, compatible == Layout(size = 16, align = 8).
-    /// // 32-bit, compatible == Layout(size = 12, align = 4)
+    /// // on 64-bit systems, compatible == Layout(size = 10, align = 8).
+    /// // 32-bit, compatible == Layout(size = 10, align = 4)
     /// ```
     #[inline]
     pub const fn to_posix_memalign_compatible(&self) -> Result<Layout, Error> {
         // first, make the alignment a multiple of `size_of::<*mut c_void>()`.
         match self.align_to_multiple_of(usize::SZ) {
             // then pad the size up to a multiple of the new alignment
-            Ok(l) => Ok(l.pad_to_align()),
+            Ok(l) => Ok(l),
             Err(_) => Err(Error::InvalidLayout(self.size(), self.align(), LayoutErr::CRoundUp))
         }
     }